requirement.txt
anyio==3.4.0
asgiref==3.4.1
awscli==1.22.26
boto3==1.20.26
botocore==1.23.26
bottle==0.12.19
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.3
contextlib2==21.6.0
contextvars==2.4
dataclasses==0.8
docutils==0.15.2
fastapi==0.70.1
feedparser==6.0.8
gensim==3.8.3
h11==0.12.0
httptools==0.3.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
jmespath==0.10.0
mecab-python==1.0.0
mecab-python3==1.0.4
numpy==1.19.5
pandas==1.1.5
pyasn1==0.4.8
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
scipy==1.5.4
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.6.0
